Run through vm.

+++ SGS/syntax.h
@@ -146,10 +146,10 @@ namespace sgs {
 		int getValue() const { return value; }
 	};
 	class FloatLiteral : public LiteralExp {
-		double value;
+		float value;
 	public:
-		explicit FloatLiteral(double value = 0) : LiteralExp(new BasicType(BT_FLOAT)), value(value) {}
-		double getValue() const { return value; }
+		explicit FloatLiteral(float value = 0) : LiteralExp(new BasicType(BT_FLOAT)), value(value) {}
+		float getValue() const { return value; }
 	};
 	class BoolLiteral : public LiteralExp {
 		bool value;
